--- packages/actions/docs/type-aliases/ValidationError.md
@@ -1,23 +1,0 @@
-[**@tevm/actions**](../README.md)
-
-***
-
-[@tevm/actions](../globals.md) / ValidationError
-
-# Type Alias: ValidationError
-
-> **ValidationError**\<\>: `object`
-
-Defined in: [packages/actions/src/Call/zCallParams.js:7](https://github.com/evmts/tevm-monorepo/blob/main/packages/actions/src/Call/zCallParams.js#L7)
-
-## Type Parameters
-
-## Type declaration
-
-### message
-
-> **message**: `string`
-
-### path?
-
-> `optional` **path**: `string`
